+++ packages/client/src/App.tsx
@@ -10,8 +10,7 @@ import MainMenuPage from './pages/main-menu-page'
 import RatingPage from './pages/rating-page'
 import SignInPage from './pages/sign-in-page'
 import SignUpPage from './pages/sign-up-page'
-import ForumPage from './pages/forum'
-import { ForumsPage } from './pages/forum/ForumsPage'
+import ForumPage from './pages/forum-page'
 import GamePage from './pages/game-page'
 import ProfilePage from './pages/profile'
 import BackgroundImage from './assets/BackgroundImage.svg'
@@ -23,19 +22,18 @@ const { initialDeckCard, cards } = imports
 
 function App() {
   useEffect(() => {
-    const fetchServerData = async () => {
-      try {
-        const serverUrl = __SERVER_URL__ || 'http://localhost:'
-        const url = `${serverUrl}${__SERVER_PORT__}`
-        const response = await fetch(url)
-        const data = await response.json()
-        console.info(data)
-      } catch (error) {
-        console.error(error)
-      }
-    }
-
-    fetchServerData()
+    // const fetchServerData = async () => {
+    //   try {
+    //     const serverUrl = __SERVER_URL__ || 'http://localhost:'
+    //     const url = `${serverUrl}${__SERVER_PORT__}`
+    //     const response = await fetch(url)
+    //     const data = await response.json()
+    //     console.info(data)
+    //   } catch (error) {
+    //     console.error(error)
+    //   }
+    // }
+    // fetchServerData()
   }, [])
 
   return (
@@ -44,7 +42,7 @@ function App() {
         <Route path={routes.login} element={<SignInPage />} />
         <Route path={routes.main} element={<MainMenuPage />} />
         <Route path={routes.registration} element={<SignUpPage />} />
-        <Route path={routes.forum} element={<>ForumPage</>} />
+        <Route path={routes.forum} element={<ForumPage />} />
         <Route path={routes.forumId} element={<>ForumTopicPage</>} />
         <Route path={routes.game} element={<GamePage />} />
         <Route path={routes.leaderboard} element={<RatingPage />} />
