--- packages/client/src/entities/forum/modalProfile/index.ts
@@ -1,1 +1,0 @@
-export { ModalProfile } from './ModalProfile'
