+++ packages/client/src/pages/forum-page/accets.ts
@@ -1,0 +1,74 @@
+export const forumTopics = [
+  {
+    id: 1,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+  {
+    id: 2,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+  {
+    id: 3,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+  {
+    id: 4,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+  {
+    id: 5,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+  {
+    id: 6,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+  {
+    id: 7,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+  {
+    id: 8,
+    title: 'Новая колода карт',
+    author: 'Иван Иванов',
+    description:
+      'Сегодня я хотел бы обсудить новую колоду карт, которая недавно вышла. Кто-нибудь уже пробовал её в игре? Как впечатления?',
+    date: '11.01.2025',
+    time: '13:36',
+  },
+]
