+++ packages/client/src/pages/forum-page/styles.module.css
@@ -1,0 +1,83 @@
+.forumPage {
+  margin: 0px 0 63px 0;
+}
+
+.forumHeaderBlock {
+  position: relative;
+  height: 108px;
+}
+
+.forumHeaderBlockScrolling {
+  border-bottom: 1px solid black;
+  box-shadow: rgba(0, 0, 0, 0.25) 0px 7px 10px;
+  transition: all 0.3s ease;
+}
+
+.forumPageTitle {
+  text-align: center;
+  color: var(--color-primary);
+  position: absolute;
+  bottom: 19px;
+  left: 50%;
+  transform: translateX(-50%);
+}
+
+.forumPageContent {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.forumPageButton {
+  margin-top: 12px;
+  height: 45px;
+}
+
+.forumPageList {
+  height: calc(100vh - 228px);
+  padding-top: 15px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  list-style-type: none;
+  overflow-y: auto;
+  scrollbar-width: none;
+}
+
+.forumPageList::-webkit-scrollbar {
+  display: none;
+  scrollbar-width: none;
+}
+
+.modalForm {
+  display: flex;
+  flex-direction: column;
+  min-width: 256px;
+  width: calc((100vw * 30) / 100);
+  gap: 30px;
+  align-items: center;
+}
+
+.modalField {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  width: 100%;
+}
+
+.modalLabel {
+  color: var(--color-background);
+  font-weight: bold;
+  font-size: 14px;
+  width: 90px;
+  text-align: end;
+}
+
+.modalProfileTextarea {
+  width: 485px;
+  height: 150px;
+  border-radius: 15px;
+  border: 1px solid var(--color-background);
+  background-color: var(--color-background-light);
+}
