--- packages/client/src/pages/forum/ForumPage.module.css
@@ -1,179 +1,0 @@
-.forumPage {
-  margin: 0px 0 63px 0;
-}
-
-.forumPageTitle {
-  text-align: center;
-  color: var(--color-primary);
-  position: absolute;
-  bottom: 19px;
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.forumHeaderBlock {
-  position: relative;
-  height: 108px;
-}
-
-.forumHeaderBlockScrolling {
-  border-bottom: 2px solid var(--color-secondary);
-  box-shadow: 0 0px 15px -2px var(--color-secondary);
-  transition: all 0.3s ease;
-}
-
-.forumPageContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.forumPageButton {
-  margin-top: 12px;
-  height: 45px;
-}
-
-.forumPageList {
-  height: calc(100vh - 228px);
-  padding-top: 15px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  list-style-type: none;
-  /* max-height: 400px; */ /* Ограничение высоты для прокрутки */
-  overflow-y: auto; /* Добавление вертикальной прокрутки */
-  scrollbar-width: none; /* Скрывает скроллбар в Firefox */
-  position: relative;
-  z-index: 1;
-}
-
-.forumPageList::-webkit-scrollbar {
-  display: none;
-  scrollbar-width: none; /* Скрывает скроллбар в Firefox */
-}
-
-.forumsContent {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  z-index: 1;
-  width: 450px;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.forumsTitle {
-  font-size: 24px;
-  font-weight: bold;
-  color: var(--color-primary);
-  width: 466px;
-}
-
-.forumsItems {
-  display: flex;
-  margin-top: 18px;
-}
-
-.forumsItemsName {
-  width: 438px;
-  height: 41px;
-  border-radius: 15px;
-  background-color: var(--color-background-light);
-  font-weight: 500;
-  font-size: 13px;
-  line-height: 13px;
-  color: var(--color-primary);
-  display: flex;
-  align-items: center;
-  padding: 0 14px;
-}
-
-.forumsItemsStats {
-  display: flex;
-}
-
-.forumsItemsTopics {
-  margin: 0 70px;
-  width: 86px;
-  height: 41px;
-  border-radius: 15px;
-  border: 2px solid var(--color-background-light);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 14px;
-  font-weight: 500;
-  font-size: 13px;
-  line-height: 13px;
-  color: var(--color-primary);
-}
-
-.forumsItemsCircle {
-  width: 30px;
-  height: 30px;
-  border-radius: 50%;
-  background-color: var(--color-background-light);
-  position: relative;
-  cursor: pointer;
-}
-
-.forumsItemsCircle:hover {
-  background-color: var(--color-secondary);
-}
-.forumsItemsCircle span {
-  display: block;
-  width: 20px;
-  height: 2px;
-  border-radius: 30px;
-  background-color: var(--color-primary);
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.forumsItemsCircle span:nth-child(2) {
-  display: block;
-  width: 20px;
-  height: 2px;
-  border-radius: 30px;
-  background-color: var(--color-primary);
-  position: absolute;
-  transform: rotate(90deg);
-  left: 5px;
-  top: 14.5px;
-}
-
-.forumsItemsReplies {
-  width: 30px;
-  height: 41px;
-  border-radius: 15px;
-  border: 2px solid var(--color-background-light);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0 14px;
-  font-weight: 500;
-  font-size: 13px;
-  line-height: 13px;
-  color: var(--color-primary);
-}
-
-.forumsTextFlex {
-  display: flex;
-  align-items: end;
-}
-
-.forumsSubtitle {
-  font-weight: bold;
-  font-size: 16px;
-  color: var(--color-secondary);
-  margin-left: 70px;
-}
-.forumsSubtitle:last-child {
-  margin-left: 140px;
-}
