--- packages/client/src/shared/arrowBackButton/index.ts
@@ -1,1 +1,0 @@
-export { ArrowBackButton } from './ArrowBackButton'
