+++ packages/client/src/shared/textarea/styles.module.css
@@ -1,0 +1,28 @@
+.textareaLabel {
+  width: 100%;
+  padding: 10px;
+  background-color: var(--color-background-light);
+  border-radius: 15px;
+  box-sizing: border-box;
+}
+
+.textarea {
+  width: 100%;
+  height: 150px;
+  background-color: inherit;
+  overflow-y: auto;
+  border: none;
+  resize: none;
+  color: var(--color-primary);
+  font-size: 14px;
+  padding: 0;
+}
+
+.textarea::-webkit-scrollbar {
+  width: 3px;
+}
+
+.textarea::-webkit-scrollbar-thumb {
+  border-radius: 10px;
+  background-color: var(--color-secondary);
+}
