+++ packages/client/src/shared/textarea/textarea.tsx
@@ -1,0 +1,9 @@
+import styles from './styles.module.css'
+
+export const Textarea = () => {
+  return (
+    <label className={styles.textareaLabel}>
+      <textarea className={styles.textarea} />
+    </label>
+  )
+}
